refactor: move AuthScreen into its own file
