+++ dataconnect/app/src/main/java/com/google/firebase/example/dataconnect/feature/profile/ProfileScreen.kt
@@ -1,32 +1,24 @@
 package com.google.firebase.example.dataconnect.feature.profile
 
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Button
 import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.google.firebase.dataconnect.movies.GetUserByIdQuery
@@ -44,7 +36,7 @@ fun ProfileScreen(
             Text((uiState as ProfileUIState.Error).errorMessage)
         }
 
-        is ProfileUIState.SignUpState -> {
+        is ProfileUIState.AuthState -> {
             AuthScreen(
                 onSignUp = { email, password, displayName ->
                     profileViewModel.signUp(email, password, displayName)
@@ -56,13 +48,13 @@ fun ProfileScreen(
         }
 
         is ProfileUIState.ProfileState -> {
-            val uiState = uiState as ProfileUIState.ProfileState
+            val ui = uiState as ProfileUIState.ProfileState
             ProfileScreen(
-                uiState.username ?: "User",
-                uiState.reviews,
-                uiState.watchedMovies,
-                uiState.favoriteMovies,
-                uiState.favoriteActors,
+                ui.username ?: "User",
+                ui.reviews,
+                ui.watchedMovies,
+                ui.favoriteMovies,
+                ui.favoriteActors,
                 onSignOut = {
                     profileViewModel.signOut()
                 }
@@ -200,81 +192,4 @@ fun FavoriteActorsList(actors: List<GetUserByIdQuery.Data.User.FavoriteActorsIte
     }
 }
 
-@Composable
-fun AuthScreen(
-    onSignUp: (email: String, password: String, displayName: String) -> Unit,
-    onSignIn: (email: String, password: String) -> Unit,
-) {
-    var isSignUp by remember { mutableStateOf(false) }
-    var email by remember { mutableStateOf("") }
-    var password by remember { mutableStateOf("") }
-    var displayName by remember { mutableStateOf("") }
 
-    Column(
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        OutlinedTextField(
-            value = email,
-            onValueChange = { email = it },
-            label = { Text("Email") }
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        OutlinedTextField(
-            value = password,
-            onValueChange = { password = it },
-            label = { Text("Password") },
-            visualTransformation = PasswordVisualTransformation()
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        if (isSignUp) {
-            OutlinedTextField(
-                value = displayName,
-                onValueChange = { displayName = it },
-                label = { Text("Name") }
-            )
-        }
-        Spacer(modifier = Modifier.height(16.dp))
-        Button(onClick = {
-            if (isSignUp) {
-                onSignUp(email, password, displayName)
-            } else {
-                onSignIn(email, password)
-            }
-        }) {
-            Text(
-                text = if (isSignUp) {
-                    "Sign up"
-                } else {
-                    "Sign in"
-                }
-            )
-        }
-//        Spacer(modifier = Modifier.height(8.dp))
-//        Button(onClick = { /* Handle Google Sign-in */ }) {
-//            Text("Sign in with Google")
-//        }
-
-        Spacer(modifier = Modifier.height(16.dp))
-        Text(
-            text = if (isSignUp) {
-                "Already have an account?"
-            } else {
-                "Don't have an account?"
-            }
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-        Button(onClick = {
-            isSignUp = !isSignUp
-        }) {
-            Text(
-                text = if (isSignUp) {
-                    "Sign in"
-                } else {
-                    "Sign up"
-                }
-            )
-        }
-    }
-}
